+++ artiq/firmware/libproto_artiq/kernel_proto.rs
@@ -85,7 +85,7 @@ pub enum Message<'a> {
     I2cBasicReply { succeeded: bool },
     I2cSwitchSelectRequest { busno: u32, address: u8, mask: u8 },
 
-    SpiSetConfigRequest { busno: u32, flags: u8, length: u8, div: u8, cs: u8 },
+    SpiSetConfigRequest { busno: u32, flags: u8, length: u8 },
     SpiWriteRequest { busno: u32, data: u32 },
     SpiReadRequest { busno: u32 },
     SpiReadReply { succeeded: bool, data: u32 },